--- src/main/java/databases_message/MessageRepoInt.java
@@ -1,19 +1,0 @@
-package databases_message;
-
-import java.util.List;
-import java.util.Map;
-
-//DsGateway
-public interface MessageRepoInt{
-
-    void save(MessageDsRequestModel requestModel);
-
-    void deleteMessage(String ID);
-
-    Map<String, Object> getMessageInfo(String messageID);
-
-    void editMessage(String ID, String content);
-
-    String getGroupMessageInfo(String GroupID);
-
-}
