+++ src/main/java/enties/MessageFactory.java
@@ -1,0 +1,12 @@
+package enties;
+
+
+public class MessageFactory{
+    public static Message create(String content, String sender, String groupID) {
+        return new Message(content,sender, groupID);
+    }
+
+
+}
+
+
