--- src/main/java/entity_message/MessageFactory.java
@@ -1,12 +1,0 @@
-package entity_message;
-
-
-public class MessageFactory{
-    public static Message create(String content, String sender, String groupID) {
-        return new Message(content,sender, groupID);
-    }
-
-
-}
-
-
