--- src/main/java/interface_adapters_message/MessageCreationFailed.java
@@ -1,8 +1,0 @@
-package interface_adapters_message;
-
-public class MessageCreationFailed extends RuntimeException {
-    public MessageCreationFailed(String error) {
-        super(error);
-    }
-
-}
