--- src/main/java/interface_adapters_message/ViewMessagePresenter.java
@@ -1,17 +1,0 @@
-package interface_adapters_message;
-
-import use_case_message_view.ViewMessageOutputboundary;
-import use_case_message_view.ViewMessageResponseModel;
-
-public class ViewMessagePresenter implements ViewMessageOutputboundary {
-
-    @Override
-    public ViewMessageResponseModel prepareSuccessView(ViewMessageResponseModel response) {
-        return response;
-    }
-
-    public ViewMessageResponseModel prepareFailView(String error) {
-        return null;
-    }
-
-}
