+++ src/main/java/message_sebd/application_business_rule/MessageInputBoundary.java
@@ -1,0 +1,8 @@
+package message_sebd.application_business_rule;
+
+public interface MessageInputBoundary {
+    MessageResponseModel create(MessageRequestModel requestModel);
+}
+
+
+
