+++ src/main/java/message_sebd/application_business_rule/MessageOutputBoundary.java
@@ -1,0 +1,7 @@
+package message_sebd.application_business_rule;
+
+public interface MessageOutputBoundary {
+    MessageResponseModel prepareSuccessView(MessageResponseModel response);
+
+    MessageResponseModel prepareFailView(String error);
+}
