+++ src/main/java/message_sebd/interface_adaptor/MessageCreationFailed.java
@@ -1,0 +1,8 @@
+package message_sebd.interface_adaptor;
+
+public class MessageCreationFailed extends RuntimeException {
+    public MessageCreationFailed(String error) {
+        super(error);
+    }
+
+}
