+++ src/main/java/message_view/application_business_rule/ViewMessageInputBoundary.java
@@ -1,0 +1,5 @@
+package message_view.application_business_rule;
+
+public interface ViewMessageInputBoundary {
+    ViewMessageResponseModel create(ViewMessageRequestModel requestModel);
+}
