+++ src/main/java/screens/GroupScreen.java
@@ -1,0 +1,13 @@
+package screens;
+
+import javax.swing.*;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+
+public class GroupScreen extends JFrame implements ActionListener{
+    @Override
+    public void actionPerformed(ActionEvent e) {
+
+    }
+    //TODO: button of all groups for the Message.
+}
