+++ src/main/java/screens/LabelTextPanel.java
@@ -1,0 +1,12 @@
+package screens;
+
+import javax.swing.*;
+
+public class LabelTextPanel extends JPanel {
+    public LabelTextPanel(JLabel label, JTextField textField) {
+        this.add(label);
+        this.add(textField);
+    }
+}
+
+//TODO: directly import from others' documents.
