--- src/main/java/screens_message/GroupScreen.java
@@ -1,13 +1,0 @@
-package screens_message;
-
-import javax.swing.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
-public class GroupScreen extends JFrame implements ActionListener{
-    @Override
-    public void actionPerformed(ActionEvent e) {
-
-    }
-    //TODO: button of all groups for the Message.
-}
