--- src/main/java/screens_message/LabelTextPanel.java
@@ -1,12 +1,0 @@
-package screens_message;
-
-import javax.swing.*;
-
-public class LabelTextPanel extends JPanel {
-    public LabelTextPanel(JLabel label, JTextField textField) {
-        this.add(label);
-        this.add(textField);
-    }
-}
-
-//TODO: directly import from others' documents.
