--- src/main/java/use_case_message_send/MessageInputBoundary.java
@@ -1,8 +1,0 @@
-package use_case_message_send;
-
-public interface MessageInputBoundary {
-    MessageResponseModel create(MessageRequestModel requestModel);
-}
-
-
-
