--- src/main/java/use_case_message_send/MessageOutputBoundary.java
@@ -1,7 +1,0 @@
-package use_case_message_send;
-
-public interface MessageOutputBoundary {
-    MessageResponseModel prepareSuccessView(MessageResponseModel response);
-
-    MessageResponseModel prepareFailView(String error);
-}
