--- src/main/java/use_case_message_view/ViewMessageInputBoundary.java
@@ -1,7 +1,0 @@
-package use_case_message_view;
-
-import java.io.IOException;
-
-public interface ViewMessageInputBoundary {
-    ViewMessageResponseModel create(ViewMessageRequestModel requestModel);
-}
