--- src/main/java/use_case_message_view/ViewMessageOutputboundary.java
@@ -1,7 +1,0 @@
-package use_case_message_view;
-
-public interface ViewMessageOutputboundary {
-    ViewMessageResponseModel prepareSuccessView(ViewMessageResponseModel response);
-
-    ViewMessageResponseModel prepareFailView(String error);
-}
