--- src/main/java/use_case_message_view/ViewMessageRequestModel.java
@@ -1,16 +1,0 @@
-package use_case_message_view;
-
-public class ViewMessageRequestModel {
-
-    private String sender;
-    private String groupID;
-
-
-    public ViewMessageRequestModel(String groupID) {
-        this.groupID = groupID;
-    }
-
-    String getGroupID(){return groupID;}
-
-    void setGroupID(String groupID){this.groupID = groupID;}
-}
